Fixed #6
